Delete category migrations from exercise + quotation
Using SQLite this migration worked but now that I've changed to PostGres on Dev and Test ENVs it fails. I suppose it actually
SHOULD fail since... there never was a category on exercise and quotation. (I added it to "Meditation" originally... but then changed the name). Maybe this is a difference between SQLite and PostGres, how it reacts to removing a column that doesn't exist