Remove unnecessary rehearsal model spec
The test I'm removing here is not needed. I want to test authorization
(can a user access another user's rehearsal) but that isn't handled on
the model level. From within the models on a Rails console level,
"anyone" can access anything. That being said, the joining of the tables
between user and rehearsal, by the simple laws of the database, it will
be impossible for another user to "find" a rehearsal not "belonging" to
him.

Authorization is a question for the controller. That's where CanCanCan
comes into play. Can I make this request - as one user, to access
another user's rehearsal? Or is this request blocked - handled another
way?