Correct jQuery to hide 'add question' button
Note that after much experimentation the cocoon:after-insert etc. events
do not work. Not sure why. In the end the key to getting anything to
work with the jQuery is that the parent element needs to be stable (non-changing, not
getting add or remove) and then you need to ALSO call the child element
in the same function. The reason is that the elements are getting dynamically generated.
See: https://elliotekj.com/2016/11/05/jquery-to-pure-js-event-listeners-on-dynamically-created-elements/